requirements.txt
# Web & HTTP
requests==2.32.3
beautifulsoup4==4.13.3
flask==3.1.3

# LINE Bot
line-bot-sdk==3.22.0

# LLM & AI
openai==2.29.0

# Embedding model
sentence-transformers==5.3.0
torch==2.5.1

# Vector database
chromadb==1.5.5

# Scheduler
schedule==1.2.2

# Environment variables
python-dotenv==1.1.0
